fix: break circular dependency and fix TypeScript compilation errors
- Extract isBotAuthor() and isAcknowledgmentComment() to shared
  comment-utils.ts, breaking the circular import between pr-monitor.ts
  and issue-conversation.ts
- Export CommentedIssueWithResponse sub-type for type-safe narrowing
  after .filter() calls (TypeScript doesn't narrow through .filter()
  without explicit type predicates)
- Fix tsc --noEmit compilation errors by using type predicates
  (i): i is CommentedIssueWithResponse in all .filter() calls
- Remove unnecessary || '' fallbacks and truthiness guards on fields
  guaranteed present by the discriminated union
- Add dashboard failure logging for issue conversation check failures
- Update dashboard fallback message to mention issue conversations
  unavailable
- Add pagination warning when issues have 100+ comments
