+++ src/core/index.ts
@@ -6,7 +6,8 @@
 export { StateManager, getStateManager, resetStateManager } from './state.js';
 export { PRMonitor, type PRCheckFailure, type FetchPRsResult, computeDisplayLabel, classifyCICheck, classifyFailingChecks } from './pr-monitor.js';
 export { IssueDiscovery, type IssueCandidate, type SearchPriority, isDocOnlyIssue, applyPerRepoCap, DOC_ONLY_LABELS } from './issue-discovery.js';
-export { IssueConversationMonitor, isAcknowledgmentComment } from './issue-conversation.js';
+export { IssueConversationMonitor } from './issue-conversation.js';
+export { isBotAuthor, isAcknowledgmentComment } from './comment-utils.js';
 export { getOctokit, checkRateLimit, type RateLimitInfo } from './github.js';
 export {
   parseGitHubUrl,
